Move endpoint_registry to public include
