+++ include/ccf/endpoint_registry.h
@@ -5,8 +5,8 @@
 #include "ccf/ds/json_schema.h"
 #include "ccf/endpoint.h"
 #include "ccf/endpoint_context.h"
+#include "ccf/rpc_context.h"
 #include "ccf/tx.h"
-#include "enclave/rpc_context.h"
 
 #include <charconv>
 #include <functional>
