+++ src/apps/js_v8/tmpl/rpc.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "enclave/rpc_context.h"
+#include "ccf/rpc_context.h"
 #include "v8.h"
 
 namespace ccf::v8_tmpl
