+++ src/consensus/aft/raft_types.h
@@ -6,7 +6,6 @@
 #include "consensus/consensus_types.h"
 #include "crypto/ecdsa.h"
 #include "ds/ring_buffer_types.h"
-#include "enclave/rpc_context.h"
 #include "enclave/rpc_handler.h"
 #include "kv/kv_types.h"
