+++ src/enclave/forwarder_types.h
@@ -2,12 +2,12 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "rpc_context.h"
-
 #include <vector>
 
 namespace enclave
 {
+  class RpcContext;
+
   class AbstractRPCResponder
   {
   public:
