+++ src/endpoints/authentication/cert_auth.cpp
@@ -3,7 +3,7 @@
 
 #include "ccf/endpoints/authentication/cert_auth.h"
 
-#include "enclave/rpc_context.h"
+#include "ccf/rpc_context.h"
 #include "service/blit.h"
 #include "service/tables/members.h"
 #include "service/tables/nodes.h"
