+++ src/endpoints/authentication/sig_auth.cpp
@@ -4,8 +4,8 @@
 #include "ccf/endpoints/authentication/sig_auth.h"
 
 #include "ccf/crypto/verifier.h"
+#include "ccf/rpc_context.h"
 #include "ds/lru.h"
-#include "enclave/rpc_context.h"
 #include "http/http_sig.h"
 #include "service/blit.h"
 #include "service/tables/members.h"
