+++ src/endpoints/json_handler.cpp
@@ -4,7 +4,7 @@
 
 #include "ccf/http_consts.h"
 #include "ccf/odata_error.h"
-#include "enclave/rpc_context.h"
+#include "ccf/rpc_context.h"
 #include "http/http_accept.h"
 #include "node/rpc/rpc_exception.h"
