+++ src/js/wrap.h
@@ -6,8 +6,8 @@
 #include "ccf/ds/logger.h"
 #include "ccf/historical_queries_interface.h"
 #include "ccf/js_plugin.h"
+#include "ccf/rpc_context.h"
 #include "ccf/tx.h"
-#include "enclave/rpc_context.h"
 #include "kv/kv_types.h"
 #include "node/network_state.h"
 #include "node/rpc/node_interface.h"
